% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddAttr.R
\name{AddAttr}
\alias{AddAttr}
\title{Add list as attributes.}
\arguments{
\item{var.any}{<any>: an object to which attributes are to be added.}

\item{attribute.lst}{<list>: a named list of new attributes.}

\item{overwrite.bln}{<logical>: whether an overwrite is required on attributes with the same name.(Default FALSE)}
}
\value{
the object with new attributes.
}
\description{
Add list as attributes to any object with or without overwrite.
}
\details{
AddAttr
}
\examples{
x <- 1:10
x
attribute.lst = list(dim=c(2,5))
x <- AddAttr(x,attribute.lst)
x
attribute.lst = list(dim=c(5,2))
x <- AddAttr(x,attribute.lst)
x
x <- AddAttr(x,attribute.lst,overwrite.bln=TRUE)
x
}
