% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindFillRows.R
\name{BindFillRows}
\alias{BindFillRows}
\title{Bind data frames after filling missing columns.}
\arguments{
\item{data.lst_df}{<data.frames or list[data.frame]>: data frames to bind or list of data.frames. If is a data.frame create a list with arguments `data.lst_df` and `...`, else `...` are ignored.}

\item{...}{<data.frames or list[data.frame]>: data frames to bind or list of data.frames.}
}
\value{
The binded data frame
}
\description{
Bind data frames by rows after filling missing columns with NA.
}
\details{
BindFillRows
}
\examples{
df1 <- data.frame(a = c(1:5), b = c(6:10))
df2 <- data.frame(a = c(11:15), b = c(16:20), c = LETTERS[1:5])
BindFillRows(df1,df2)
BindFillRows(list(df1,df2))
}
