% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakVector.R
\name{BreakVector}
\alias{BreakVector}
\title{Cut a vector.}
\arguments{
\item{x.num}{<numeric>: numerical vector.}

\item{min.num}{<numeric>: minimal fixed value.}

\item{center.num}{<numeric>: center fixed value.}

\item{max.num}{<numeric>: maximal fixed value.}

\item{n.num}{<numeric>: number of tile (return n.num+1 breaks).}

\item{method.chr}{<character>: kind of breaking. "linear" or "density". (Default "linear")}
}
\value{
Numerical vector of breaks.
}
\description{
Compute the n+1 breaks of a vector in a linear or density based way with the possibility to fix minimal, center and maximal values.
}
\details{
BreakVector
}
\examples{
set.seed(31415)
x.num <- rnorm(100,50,200)
BreakVector(x.num=x.num, n.num=9)
}
