% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CbrtNorm.R
\name{CbrtNorm}
\alias{CbrtNorm}
\title{Cubic root normalization.}
\arguments{
\item{x.num}{<numeric>: numerical vector.}

\item{a.num}{<numeric>: constant value to add before compute cubic root. (Default 0)}
}
\value{
normalized vector.
}
\description{
Normalize data by cubic root and adding a constant value.
}
\details{
CbrtNorm
}
\examples{
set.seed(542972)
x.num <- rnbinom(1000, 1, 0.01)
plot(density(x.num))
plot(density(CbrtNorm(x.num, abs(min(x.num)))))
}
