% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col2Hsl.R
\name{Col2Hsl}
\alias{Col2Hsl}
\title{Convert a color in HSL format.}
\arguments{
\item{color.col}{<charcater>: a vector of the color's present in base R colors().}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
A vector of the color's hsl code.
}
\description{
Convert a color in HSL (Hue,Saturation,Light) format.
}
\details{
Col2Hsl
}
\examples{
Col2Hsl("#ebb0107f",alpha.bln=TRUE)
}
