% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col2Rgb.R
\name{Col2Rgb}
\alias{Col2Rgb}
\title{Convert a color in to RGB format.}
\arguments{
\item{color.col}{<charcater>: a vector of the color's present in base R colors().}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's RGB code.
}
\description{
Convert a color in to RGB format.
}
\details{
Col2Rgb
}
\examples{
Col2Rgb(c("green","red"),alpha.bln=TRUE)
}
