% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colorize.R
\name{Colorize}
\alias{Colorize}
\title{Fixe saturation at 100\%.}
\arguments{
\item{color.col}{<character or numeric>: color in any format.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's rgb code.
}
\description{
Fixe saturation at 100\%.
}
\details{
Colorize
}
\examples{
Colorize("#ebb0107f",alpha.bln=TRUE)
}
