% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertColorFormat.R
\name{ConvertColorFormat}
\alias{ConvertColorFormat}
\title{Convert a color format.}
\arguments{
\item{color.col}{<charcater>: a color in any format.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}

\item{format.str}{<character>: a character (case insensitive) that give the format you want ("Hex", "Hsl", "Rgb"). (Default "Hex")}
}
\value{
An integer vector of the color's rgb code.
}
\description{
Convert a color in a desired format.
}
\details{
ConvertColorFormat
}
\examples{
ConvertColorFormat("#ebb0107f", alpha.bln=TRUE, format.str="Rgb")
ConvertColorFormat("#ebb0107f", alpha.bln=TRUE, format.str="Hsl")
ConvertColorFormat(c(235,176,16,127), alpha.bln=TRUE, format.str="Hex")
ConvertColorFormat(c(235,176,16,127), alpha.bln=TRUE, format.str="Hsl")
ConvertColorFormat(c(43.8,0.873,0.492,0.498), alpha.bln=TRUE, format.str="Rgb")
ConvertColorFormat(c(43.8,0.873,0.492,0.498), alpha.bln=TRUE, format.str="Hex")
}
