% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darken.R
\name{Darken}
\alias{Darken}
\title{Fixe light at 25\%.}
\arguments{
\item{color.col}{<character or numeric>: color in any format}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's rgb code.
}
\description{
Fixe light at 25\%.
}
\details{
Darken
}
\examples{
Darken("#ebb0107f",alpha.bln=TRUE)
}
