% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindElbow.R
\name{FindElbow}
\alias{FindElbow}
\title{Find the elbow point on a curve.}
\arguments{
\item{x.num}{<numeric>: numeric vector of x coordinates of the curve.}

\item{y.num}{<numeric>: numeric vector of y coordiantes of the curve.}
}
\value{
Numeric vector of elbow coordinates.
}
\description{
Find the elbow point on a curve.
}
\details{
FindElbow
}
\examples{
x.num = 1:10
y.num = 1/sqrt(x.num)
elbow = FindElbow(x.num, y.num)
elbow
plot(y.num~x.num,type='l')
points(x=elbow[1],y=elbow[2])
}
