% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussBox.R
\name{GaussBox}
\alias{GaussBox}
\title{One dimension Gaussian kernel.}
\arguments{
\item{sd.num}{<numeric>: standard deviation parameter of the gaussian. (Default 1)}

\item{boxSize.num}{<numeric>: size of kernel. If NULL size is 1+4*sd.num (Default NULL)}

\item{scale.chr}{<character>: scaling kind of box. If "1" sum of kernel equal 1. If "int" Minimal value of kernel is 1 and all entry are integer. If "none", kernel is not scale. (Default "1")}
}
\value{
numerical vector.
}
\description{
One dimension Gaussian kernel.
}
\details{
GaussBox
}
\examples{
GaussBox( scale.chr="none")
GaussBox( scale.chr="1")
sum(GaussBox( scale.chr="1"))
GaussBox( scale.chr="int")
}
