% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussKernel.R
\name{GaussKernel}
\alias{GaussKernel}
\title{Two dimension Gaussian kernel.}
\arguments{
\item{sd.num}{<numeric>: standard deviation parameter of the gaussian. (Default 1)}

\item{kernSize.num}{<numeric>: size of kernel. If null size is 1+4*sd.num. (Default NULL)}

\item{scale.chr}{<character>: scaling kind of box. If "1" sum of kernel equal 1. If "int" Minimal value of kernel is 1 and all entry are integer. If "none", kernel is not scale. (Default "1")}

\item{by.num}{<numerical>: step size for compute Gaussian function in two dimension (Default 0.01).}
}
\value{
numerical matrix.
}
\description{
Two dimension Gaussian kernel.
}
\details{
GaussKernel
}
\examples{
GaussKernel(scale.chr=c("none"))
GaussKernel(scale.chr=c("1"))
sum(GaussKernel(scale.chr=c("1")))
GaussKernel(scale.chr=c("int"))
}
