% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetColorFormat.R
\name{GetColorFormat}
\alias{GetColorFormat}
\title{Get format of a color}
\arguments{
\item{color.col}{<charcater or numerical>: a color in any format.}
}
\value{
A character of color format.
}
\description{
Return the format of a color.
}
\details{
GetColorFormat
}
\examples{
GetColorFormat("#ebb0107f")
GetColorFormat(c(235,176,16,127))
GetColorFormat(c(43.8,0.873,0.492,0.498))
GetColorFormat("green")
}
