% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hex2Rgb.R
\name{Hex2Rgb}
\alias{Hex2Rgb}
\title{Convert a color in hexadecimal to RGB.}
\arguments{
\item{hex.col}{<charcater>: a character of the color's hexadecimal code.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's RGB code.
}
\description{
Convert a color in hexadecimal format to RGB format.
}
\details{
Hex2Rgb
}
\examples{
Hex2Rgb("#ebb0107f",alpha.bln=TRUE)
}
