% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hsl2Hex.R
\name{Hsl2Hex}
\alias{Hsl2Hex}
\title{Convert a color in HSL to hexadecimal.}
\arguments{
\item{hsl.col}{<charcater>: a vector of the color's HSL code.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
A character of the color's hexadecimal code.
}
\description{
Convert a color in HSL (Hue,Saturation,Light) to hexadecimal format.
}
\details{
Hsl2Hex
}
\examples{
Hsl2Hex(c(43.8,0.873,0.492,0.498),alpha.bln=TRUE)
}
