% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IqrThreshold.R
\name{IqrThreshold}
\alias{IqrThreshold}
\title{Find threshold for outliers triming based on inter quartile range.}
\arguments{
\item{x.num}{<numeric>: numeric vector.}

\item{iqrFactor.num}{<numeric>: inter quartile range factor. (Default 1.5)}

\item{bounds.chr}{<character>: bounds to return, "lower", "upper" or "both". (Default "both")}
}
\value{
numerical vector of minmal and/or maximal thresholds values for outliers triming.
}
\description{
Find threshold for outliers triming based on inter quartile range.
}
\details{
IqrThreshold
}
\examples{
set.seed(1111)
x.num <- rnorm(1000)
x.num <- sort(x.num)
x.num
IqrThreshold(x.num, bounds.chr="lower")
IqrThreshold(x.num, bounds.chr="both")
IqrThreshold(x.num, bounds.chr="upper")
}
