% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsHex.R
\name{IsHex}
\alias{IsHex}
\title{Check if a color is in hexadecimal color format.}
\arguments{
\item{color.col}{<character or numeric>: a color.}
}
\value{
a logical.
}
\description{
Check if a color is in hexadecimal color format.
}
\details{
IsHex
}
\examples{
IsHex("red")
IsHex("#FFFFFF")
IsHex(c(125,125,125))
IsHex(c(43.8,0.873,0.492))
}
