% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsPath.R
\name{IsPath}
\alias{IsPath}
\title{Check if a character string is a file path.}
\arguments{
\item{var.any}{<any>: a variable}
}
\value{
A boolean
}
\description{
Check if a character string is a path of an existing file.
}
\details{
IsPath
}
\examples{
file.create("test.file", showWarnings = TRUE)
IsPath("test.file")
IsPath("test.file2")
unlink("test.file")
}
