% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log2Norm.R
\name{Log2Norm}
\alias{Log2Norm}
\title{Log2 normalization.}
\arguments{
\item{x.num}{<numeric>: numerical vector.}

\item{a.num}{<numeric>: constant value to add before compute cubic root. (Default 0)}
}
\value{
normalized vector.
}
\description{
Normalize data by logarithme (base 2) and adding a constant value.
}
\details{
Log2Norm
}
\examples{
set.seed(542972)
x.num <- rnbinom(1000, 1, 0.01)
plot(density(x.num))
plot(density(Log2Norm(x.num, abs(min(x.num)))))
}
