% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianSkewness.R
\name{MedianSkewness}
\alias{MedianSkewness}
\title{Pearson median skewness.}
\arguments{
\item{x.num}{<numeric>: numerical vector.}
}
\value{
skewness value.
}
\description{
Pearson median skewness.
}
\details{
MedianSkewness
}
\examples{
set.seed(542972)
MedianSkewness(rnbinom(10000, 1, 0.01))
set.seed(542972)
MedianSkewness(rnorm(10000, 1, 0.01))
}
