% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{Normalize}
\alias{Normalize}
\title{Normalize data.}
\arguments{
\item{x.num}{<numeric>: numerical vector}

\item{a.num}{<numeric>: constant value to add before compute normalisation too avoid negativ number. If null auto dertermine a.num. (Default NULL)}
}
\value{
Normalized vector
}
\description{
Normalize data by with square root, cubic root or log2 based on minimal skewness.
}
\details{
Normalize
}
\examples{
set.seed(542972)
x.num <- rnbinom(1000, 1, 0.01)
plot(density(x.num))
plot(density(Normalize(x.num)))
}
