% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PadMtx.R
\name{PadMtx}
\alias{PadMtx}
\title{Add a value around a matrix.}
\arguments{
\item{mat.mtx}{<matrix>: numerical matrix.}

\item{padSize.num}{<numeric>: number of columns or rows to add. (Default 1)}

\item{value.num}{<numeric>: value to add. If Null create mirror of choosen sides. (Default 0)}

\item{side.chr}{<character>: side to pad, must be one or some of 'top','bot','right' or 'left'. (Default c('top','bot','right','left') )}
}
\value{
a matrix.
}
\description{
Add a value around a matrix.
}
\details{
PadMtx
}
\examples{
mat.mtx = matrix(1:25,5,5)
PadMtx(mat.mtx=mat.mtx, padSize.num=1, value.num=0, side.chr=c('top','bot','right','left') )
PadMtx(mat.mtx=mat.mtx, padSize.num=1, value.num=NULL, side.chr=c('top','bot','right','left') )
PadMtx(mat.mtx=mat.mtx, padSize.num=1, value.num=0, side.chr=c('right','left') )
PadMtx(mat.mtx=mat.mtx, padSize.num=1, value.num=0, side.chr=c('top') )
}
