% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pastelize.R
\name{Pastelize}
\alias{Pastelize}
\title{Fixe light at 75\%.}
\arguments{
\item{color.col}{<character or numeric>: color in any format.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's rgb code.
}
\description{
Fixe light at 75\%.
}
\details{
Pastelize
}
\examples{
Pastelize("#ebb0107f",alpha.bln=TRUE)
}
