% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceRun.R
\name{ReduceRun}
\alias{ReduceRun}
\title{Apply a function over two RLE}
\arguments{
\item{first.rle}{<rle or Rle>: first rle.}

\item{second.rle}{<rle or Rle>>: second rle.}

\item{reduceFun.chr}{<character>: name of a function to apply e.g paste, sum, mean.}

\item{...}{<...>: Other parameter for the reduce function.}
}
\value{
reduced Rle
}
\description{
Apply a function on the values over two RLE and return one RLE.
}
\details{
ReduceRun
}
\examples{
first.rle = rle(c("A","A","B"))
second.rle = rle(c("A","B","B"))
ReduceRun(first.rle=first.rle, second.rle=second.rle, reduceFun.chr="paste", sep="_" )
first.rle = S4Vectors::Rle(c(1,2,3))
second.rle = S4Vectors::Rle(c(5,5,5))
ReduceRun(first.rle=first.rle, second.rle=second.rle, reduceFun.chr="sum")
}
