% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResizeMatrix.R
\name{ResizeMatrix}
\alias{ResizeMatrix}
\title{Resize a matrix}
\arguments{
\item{matrice.mtx}{<matrix>: a numericam matrix to resize.}

\item{newDim.num}{<integer>: the number of rows and cols in resized matrix.}
}
\value{
resized matrix.
}
\description{
Resize a numericam matrix in new dimension.
}
\details{
ResizeMatrix
}
\examples{
matrice.mtx <- matrix(0,11,11)
matrice.mtx[which(as.logical(1:(11*11)\%\%2))] <- 1:ceiling((11*11)/2)
matrice.mtx[2,] <- 100
matrice.mtx[,7] <- 200
matrice.mtx
ResizeMatrix(matrice.mtx=matrice.mtx, newDim.num=c(7,7))
ResizeMatrix(matrice.mtx=matrice.mtx, newDim.num=c(13,13))
}
