% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgb2Hsl.R
\name{Rgb2Hsl}
\alias{Rgb2Hsl}
\title{Convert a color in RGB to HSL format.}
\arguments{
\item{rgb.col}{<integer>: an integer of the color's RGB code.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
A vector of the color's HSL code.
}
\description{
Convert a color in RGB format to HSL (Hue,Saturation,Light) format.
}
\details{
Rgb2Hsl
}
\examples{
Rgb2Hsl(c(235,176,16,127),alpha.bln=TRUE)
}
