% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocket.R
\name{rocket}
\alias{rocket}
\title{Create a rocket palette}
\arguments{
\item{paletteLength.num}{<numeric>: color number.}

\item{space.chr}{<numeric>: a character string; interpolation in RGB or CIE Lab color spaces. See ?grDevices::colorRamp for more details. (Default "rgb")}

\item{interpolate.chr}{<numeric>: use spline or linear interpolation. See ?grDevices::colorRamp for more details. (Default "linear")}

\item{bias.num}{<numeric>: a positive number. Higher values give more widely spaced colors at the high end. See ?grDevices::colorRamp for more details. (Default 1)}
}
\value{
A vector of color.
}
\description{
Create a rocket palette.
}
\details{
rocket
}
\examples{
rocket(9)
ggPalette(rocket(255))
}
