% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtNorm.R
\name{SqrtNorm}
\alias{SqrtNorm}
\title{Square root normalization.}
\arguments{
\item{x.num}{<numeric>: Numerical vector.}

\item{a.num}{<numeric>: Constant value to add before compute square root. (Default 0)}
}
\value{
Normalized vector.
}
\description{
Normalize data by square root and adding a constant value.
}
\details{
SqrtNorm.
}
\examples{
set.seed(542972)
x.num <- rnbinom(1000, 1, 0.01)
plot(density(x.num))
plot(density(SqrtNorm(x.num, abs(min(x.num)))))
}
