% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tarnished.R
\name{Tarnished}
\alias{Tarnished}
\title{Fixe saturation at 25\%.}
\arguments{
\item{color.col}{<character or numeric>: color in any format.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's rgb code.
}
\description{
Fixe saturation at 25\%.
}
\details{
Tarnished
}
\examples{
Tarnished("#ebb0107f",alpha.bln=TRUE)
}
