% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitLengthScale.R
\name{UnitLengthScale}
\alias{UnitLengthScale}
\title{Unit length scaling.}
\arguments{
\item{x.num}{<numeric>: numerical vector.}
}
\value{
scaled numeric vector.
}
\description{
Scale values with unit lenght method.
}
\details{
UnitLengthScale
}
\examples{
set.seed(655213)
x.num <- rnorm(500,500)
set.seed(522613)
y.num <- rnorm(500,100)
plot(density(x.num),col="red",xlim=c(min(y.num),max(x.num )))
lines(density(y.num),col="green")
plot(density(UnitLengthScale(x.num)),
col="red",
xlim=c(
min(UnitLengthScale(y.num)),
max(UnitLengthScale(x.num) )))
lines(density(UnitLengthScale(y.num)),col="green")
}
